add day2

+++ template/day01.cpp
@@ -8,13 +8,14 @@
 #include <algorithm>
 #include <filesystem>
 #include <numeric>
+#include <cstdint>
 
 #include "util.hpp"
 
 using namespace std;
 
 
-int d01pt1(const filesystem::path& file_path) {
+int64_t d01pt1(const filesystem::path& file_path) {
     cout << endl << "Day 01 - Pt1: " << file_path.filename() << endl;
     auto lines = read_lines(file_path);
     int result = 0;
@@ -25,7 +26,7 @@ int d01pt1(const filesystem::path& file_path) {
     return result;
 }
 
-int d01pt2(const filesystem::path& file_path) {
+int64_t d01pt2(const filesystem::path& file_path) {
     cout << endl << "Day 01 - Pt2: " << file_path.filename() << endl;
     auto lines = read_lines(file_path);
     int result = 0;
